+++ pyffttools/stretchbuilder.py
@@ -22,4 +22,4 @@ class StretchBuilder(BuilderBase):
         variance2 = np.var(data2)
         a = math.sqrt(variance/variance2)
         b = average - average2*a
-        return a*data2+b
+        return (a*data2+b).astype(data.dtype)
